Refactored statement generation to prepare for typst support
