+++ task-maker-format/src/ioi/statement/booklet.rs
@@ -1,21 +1,13 @@
-use std::collections::HashSet;
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
+use std::path::PathBuf;
 
-use anyhow::{anyhow, Context, Error};
-use askama::Template;
-use itertools::Itertools;
-use regex::Regex;
+use anyhow::{anyhow, bail, Context, Error};
 use serde::{Deserialize, Serialize};
 use typescript_definitions::TypeScriptify;
 
-use task_maker_dag::{Execution, ExecutionCommand, File};
-use task_maker_diagnostics::Diagnostic;
+use crate::ioi::statement::statement::{Language, Statement};
+use crate::EvaluationData;
 
-use crate::ioi::statement::statement::Statement;
-use crate::ioi::BOOKLET_PRIORITY;
-use crate::ui::UIMessageSender;
-use crate::{bind_exec_callbacks, ui::UIMessage, EvaluationData, Tag, UISender, DATA_DIR};
+use super::get_language_from_extension;
 
 /// Configuration of a `Booklet`, including the setting from the contest configuration.
 #[derive(Debug, Clone, Serialize, Deserialize, Default, TypeScriptify)]
@@ -42,24 +34,6 @@ pub struct BookletConfig {
     pub intro_page: Option<PathBuf>,
 }
 
-/// Template to use to render the `booklet.tex` file.
-#[derive(Template)]
-#[template(path = "booklet.tex", escape = "none", syntax = "tex")]
-pub struct BookletTemplate {
-    language: String,
-    show_solutions: String,
-    show_summary: String,
-    font_enc: String,
-    input_enc: String,
-    description: String,
-    location: String,
-    date: String,
-    logo: String,
-    packages: String,
-    tasks: String,
-    intro_page: String,
-}
-
 /// A `Booklet` is a pdf file containing the statements of some tasks. It is compiled from a series
 /// of `.tex` files defined by `Statement` objects. The compiled pdf file is then copied somewhere.
 #[derive(Debug, Clone, Serialize, Deserialize, TypeScriptify)]
@@ -70,6 +44,8 @@ pub struct Booklet {
     pub statements: Vec<Statement>,
     /// Where to copy the booklet.
     pub dest: PathBuf,
+    /// The language extension of the statements
+    pub lang: Option<String>,
 }
 
 /// Part of the schema of `contest.yaml`, used for extracting the configuration of the booklet.
@@ -98,12 +74,29 @@ impl Booklet {
             config,
             dest: dest.into(),
             statements: Vec::new(),
+            lang: None,
         }
     }
 
     /// Add a `Statement` to this booklet.
-    pub fn add_statement(&mut self, statement: Statement) {
+    pub fn add_statement(&mut self, statement: Statement) -> Result<(), Error> {
+        let Some(statement_lang) = statement.path.extension() else {
+            bail!("Statement at {} has no extension", statement.path.display())
+        };
+        let statement_lang = statement_lang.to_string_lossy().to_string();
+
+        if self
+            .lang
+            .as_ref()
+            .is_some_and(|lang| *lang != statement_lang)
+        {
+            bail!("Booklet contains statements with two or more different extensions");
+        }
+
+        self.lang = Some(statement_lang);
+
         self.statements.push(statement);
+        Ok(())
     }
 
     /// Build the booklet, eventually coping the final PDF to the specified destination.
@@ -114,203 +107,14 @@ impl Booklet {
             .ok_or_else(|| anyhow!("Invalid destination file {}", self.dest.display()))?
             .to_string_lossy()
             .to_string();
-        let mut exec = Execution::new(
-            "Compilation of the booklet",
-            ExecutionCommand::system("latexmk"),
-        );
-        exec.args(vec![
-            "-shell-escape",
-            "-f",
-            "-interaction=nonstopmode",
-            "-pdf",
-            "booklet.tex",
-        ]);
-        exec.limits_mut()
-            .read_only(false)
-            .allow_multiprocess()
-            .add_extra_readable_dir("/etc")
-            .mount_tmpfs(true);
-        exec.tag(Tag::Booklet.into());
-        exec.priority(BOOKLET_PRIORITY);
-        let output = exec.output("booklet.pdf");
 
-        let source = File::new("Source of the booklet");
-        let tex = self.make_tex();
-        exec.input(&source, "booklet.tex", false);
-        eval.dag.provide_content(source, tex.into_bytes());
+        // If no statement has been added to booklet, there is nothing to do
+        let Some(lang) = &self.lang else {
+            return Ok(());
+        };
+        let builder = get_language_from_extension(lang)?;
 
-        for statement in self.statements.iter() {
-            let name = &statement.config().name;
-            let tex = File::new(format!("Source of statement of {}", name));
-            exec.input(&tex, Path::new(&name).join("statement.tex"), false);
-            eval.dag.provide_content(tex, statement.tex().into_bytes());
-            let base_dir = PathBuf::from(&name);
-            let deps = statement
-                .build_deps(eval, &booklet_name, &self.config)
-                .context("Failed to build booklet dependencies")?;
-
-            for (path, file) in &deps {
-                exec.input(file, base_dir.join(path), false);
-            }
-
-            let deps_paths: Vec<_> = deps.iter().map(|(path, _file)| path.as_path()).collect();
-
-            // provide defaults limiti.py and constraints.py if needed
-            if !deps_paths.contains(&Path::new("limiti.py"))
-                && deps_paths.contains(&Path::new("limiti.yaml"))
-            {
-                let path = DATA_DIR.join("statements/limiti.py");
-
-                let file = File::new("Default limiti.py");
-                exec.input(&file, base_dir.join("limiti.py"), false);
-                eval.dag.provide_file(file, &path)?;
-            }
-
-            if !deps_paths.contains(&Path::new("constraints.py"))
-                && deps_paths.contains(&Path::new("constraints.yaml"))
-            {
-                let path = DATA_DIR.join("statements/constraints.py");
-
-                let file = File::new("Default constraints.py");
-                exec.input(&file, base_dir.join("constraints.py"), false);
-                eval.dag.provide_file(file, &path)?;
-            }
-        }
-
-        // copy all the files from the data/statements directory
-        let data_dir = DATA_DIR.join("statements");
-        let glob_pattern = data_dir.to_string_lossy().to_string() + "/**/*";
-        for path in glob::glob(&glob_pattern).context("Invalid glob pattern")? {
-            let path = path.context("Failed to iterate with glob")?;
-            if !path.is_file() {
-                continue;
-            }
-            let file = File::new(format!(
-                "Booklet template file {:?}",
-                path.file_name().context("Invalid template file")?
-            ));
-            eval.dag
-                .provide_file(file.clone(), &path)
-                .context("Failed to provide statement file")?;
-            exec.input(file, path.strip_prefix(&data_dir)?, false);
-        }
-
-        bind_exec_callbacks!(
-            eval,
-            exec.uuid,
-            |status, name| UIMessage::IOIBooklet { name, status },
-            booklet_name
-        )?;
-        if eval.dag.data.config.copy_logs {
-            let log_dir = eval.task_root.join("bin/logs/booklets");
-            let stderr_dest = log_dir.join(format!("{}.stderr.log", booklet_name));
-            let stdout_dest = log_dir.join(format!("{}.stdout.log", booklet_name));
-            eval.dag
-                .write_file_to_allow_fail(exec.stderr(), stderr_dest, false);
-            eval.dag
-                .write_file_to_allow_fail(exec.stdout(), stdout_dest, false);
-        }
-        let sender = eval.sender.clone();
-        exec.capture_stdout(1024 * 1024 * 1024);
-        let dest = self.dest.file_name().unwrap().to_owned();
-        eval.dag.on_execution_done(&exec.uuid, move |res| {
-            if let Some(content) = &res.stdout {
-                Booklet::emit_warnings(dest, content, sender)?;
-            }
-            Ok(())
-        });
-        eval.dag.add_execution(exec);
-        // latexmk may fail but still produce a good-enough pdf file
-        eval.dag.write_file_to_allow_fail(output, &self.dest, false);
-
-        Ok(())
-    }
-
-    /// Build the main booklet.tex source file by combining the info from all the statements and
-    /// expanding the template.
-    fn make_tex(&self) -> String {
-        let mut packages = HashSet::new();
-        let mut tasks = Vec::new();
-        for statement in self.statements.iter() {
-            for package in statement.packages() {
-                packages.insert(package);
-            }
-            tasks.push(format!(
-                r"\subimport{{./{}/}}{{statement.tex}}",
-                statement.config().name
-            ));
-        }
-        BookletTemplate {
-            language: self.config.language.clone(),
-            show_solutions: Booklet::bool_to_tpl_string(
-                self.config.show_solutions,
-                "showsolutions",
-            ),
-            show_summary: Booklet::bool_to_tpl_string(self.config.show_summary, "showsummary"),
-            font_enc: self.config.font_enc.clone(),
-            input_enc: self.config.input_enc.clone(),
-            description: self.config.description.clone().unwrap_or_default(),
-            location: self.config.location.clone().unwrap_or_default(),
-            date: self.config.date.clone().unwrap_or_default(),
-            logo: self.config.logo.clone().unwrap_or_default(),
-            packages: packages.iter().sorted().join("\n"),
-            tasks: tasks.join("\n"),
-            intro_page: self
-                .config
-                .intro_page
-                .clone()
-                .map(std::fs::read_to_string)
-                .unwrap_or_else(|| Ok(String::new()))
-                .unwrap_or_default(),
-        }
-        .to_string()
-    }
-
-    /// Return a string which is `if_true` if `b` is true, otherwise an empty string.
-    fn bool_to_tpl_string(b: bool, if_true: &str) -> String {
-        if b { if_true } else { "" }.to_string()
-    }
-
-    /// Given the content of the log from latexmk, extract the errors and emit them as warnings.
-    fn emit_warnings(
-        booklet_name: impl AsRef<Path>,
-        content: &[u8],
-        sender: Arc<Mutex<UIMessageSender>>,
-    ) -> Result<(), Error> {
-        lazy_static! {
-            static ref FIND_ERRORS: Regex =
-                Regex::new(r"(?ms)^!(?: LaTeX Error:)? ([^\n]+).*?(^l\.\d+)")
-                    .expect("Invalid regex");
-        }
-        // latexmk sometimes emit the same warning more than once
-        let mut errors = HashSet::new();
-        for cap in FIND_ERRORS.captures_iter(&String::from_utf8_lossy(content)) {
-            let line = cap[2]
-                .strip_prefix("l.")
-                .and_then(|line| line.parse::<i32>().ok());
-            errors.insert((line, cap[1].to_string()));
-        }
-        if !errors.is_empty() {
-            let note = errors
-                .into_iter()
-                .sorted()
-                .map(|(line, error)| {
-                    if let Some(line) = line {
-                        format!("Line {}: {}", line, error)
-                    } else {
-                        error
-                    }
-                })
-                .join("\n");
-            sender.add_diagnostic(
-                Diagnostic::warning(format!(
-                    "Found Latex errors while compiling the booklet {}",
-                    booklet_name.as_ref().display()
-                ))
-                .with_note(note),
-            )?;
-        }
-        Ok(())
+        builder.create_execution(self, booklet_name, eval)
     }
 }
 
@@ -376,31 +180,36 @@ impl BookletConfig {
 
 #[cfg(test)]
 mod tests {
+    use std::path::Path;
+
     use crate::ioi::StatementConfig;
 
     use super::*;
 
-    fn get_outputs_with_logs(task_root: &Path, copy_logs: bool) -> Vec<PathBuf> {
+    fn get_outputs_with_logs(task_root: &Path, copy_logs: bool) -> Result<Vec<PathBuf>, Error> {
         let (mut eval, _recv) = EvaluationData::new(task_root);
         eval.dag.data.config.copy_logs = copy_logs;
         let mut booklet = Booklet::new(BookletConfig::default(), task_root.join("dest.pdf"));
         std::fs::write(task_root.join("text.tex"), "loltex").unwrap();
         let statement =
             Statement::new(task_root.join("text.tex"), StatementConfig::default()).unwrap();
-        booklet.add_statement(statement);
+        booklet.add_statement(statement)?;
         booklet.build(&mut eval).unwrap();
-        eval.dag
+        let ret = eval
+            .dag
             .file_callbacks()
             .values()
             .filter_map(|f| f.write_to.as_ref())
             .map(|f| f.dest.clone())
-            .collect()
+            .collect();
+        Ok(ret)
     }
 
     #[test]
     fn test_logs_emitted_with_copy_logs() {
         let tmpdir = tempfile::TempDir::new().unwrap();
-        let outputs = get_outputs_with_logs(tmpdir.path(), true);
+        let outputs = get_outputs_with_logs(tmpdir.path(), true)
+            .expect("Incorrectly found mismatching extensions");
         let stderr_path = tmpdir.path().join("bin/logs/booklets/dest.pdf.stderr.log");
         let stdout_path = tmpdir.path().join("bin/logs/booklets/dest.pdf.stdout.log");
         assert!(outputs.contains(&stderr_path));
@@ -410,7 +219,8 @@ mod tests {
     #[test]
     fn test_logs_not_emitted_by_default() {
         let tmpdir = tempfile::TempDir::new().unwrap();
-        let outputs = get_outputs_with_logs(tmpdir.path(), false);
+        let outputs = get_outputs_with_logs(tmpdir.path(), false)
+            .expect("Incorrectly found mismatching extensions");
         let stderr_path = tmpdir.path().join("bin/logs/booklets/dest.pdf.stderr.log");
         let stdout_path = tmpdir.path().join("bin/logs/booklets/dest.pdf.stdout.log");
         assert!(!outputs.contains(&stderr_path));
